+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/groups/components/org_sources_modal.test.tsx
@@ -13,7 +13,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 
 import { GroupManagerModal } from './group_manager_modal';
-import { OrgSourcesModal } from './shared_sources_modal';
+import { OrgSourcesModal } from './org_sources_modal';
 import { SourcesList } from './sources_list';
 
 const group = groups[0];
