+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/groups/components/org_sources_modal.tsx
